feat: Show nicer error if cachedir can't be created (#88)
If we cannot create the cache directory to install the elk.js
dependencies, we used to crash with whatever error message the OS gave
us. This can sometimes be misleading, e.g. when the message is talking
about missing permissions, but in reality `$HOME` or `$XDG_CACHE_HOME`
is simply set to a broken value.

This commit catches an OSError here and instead shows a nicer error
message (and also much more concise stack trace) with a concrete hint
towards `$HOME`.

This is mostly relevant when running inside a container or as a system
service, as it's much easier to accidentally miss some envvars there.