fix: Omit exchange item element labels
Because of elkjs limitations edge labels cannot be displayed properly at this point without sacrificing the desired layout.
Edge labels will be omitted completely as to not create confusion with them showing up on the top left corner.
They should be added back when the issues with the edge labels got fixed.