fix(elkjs): Download last full release during X.devY versions
If deno isn't available locally, the elkjs manager instance tries to
download and use a runtime with the exact version of context-diagrams.
This works fine while sitting exactly on a release tag, but for all
other cases, setuptools-scm will generate intermediate version numbers
like `1.0.1.dev5+ABCD123`. These obviously do not correspond to releases
on Github, nor do they exist as binaries on disk, which is why
context-diagrams would always fail to spawn the layouting server.

This commit changes the version detection logic to step back to the
latest release tag if such an intermediate dev version was generated.